aime_epic1/backend/django/requirements.txt
Django>=5.0.6
mysqlclient>=2.2.4
python-dotenv>=1.0.1
gunicorn==22.0.0
django-environ==0.11.2
whitenoise==6.6.0
requests==2.32.3
